Added multi-line resusable test file
